authentication added

+++ resources/views/layouts/page.blade.php
@@ -18,15 +18,28 @@
             <ul class="nav navbar-nav">
                 <li class="active"><a href="/">Home</a></li>
                 <li><a href="/about">About</a></li>
-                <li><a href="/register">Register</a></li>
-                <li><a href="/login">Login</a></li>
+
+                @if(Auth::check())
+                    <li><a href="/logout">Logout</a></li>
+
+                @else
+
+                    <li><a href="/register">Register</a></li>
+
+                    <li><a href="/login">Login</a></li>
+                @endif
                 <li><a class="dropdown-toggle" data-toggle="dropdown" href="#">Articles
                         <span class="caret"></span></a>
                     <ul class="dropdown-menu">
-                        <li><a href="/add_article">Add article</a></li>
-                        <li><a href="/edit_articles">Edit article</a></li>
-                        <li><a href="/delete_article">Delete article</a></li>
-                        <li><a href="/articles">All articles</a></li>
+                       @if(Auth::check())
+                            <li><a href="/add_article">Add article</a></li>
+                            <li><a href="/edit_articles">Edit article</a></li>
+                            <li><a href="/delete_article">Delete article</a></li>
+                            <li><a href="/articles">All articles</a></li>
+                            @else
+                            <li><a href="/articles">All articles</a></li>
+                            @endif
+
                     </ul>
                 </li>
